fix: disable `esModule` output for css-loader
From [1] we are told that the esModule interop was added in
vue-style-loader@4.1.3. It did work well with css-loader@5.
However, since we've upgraded our css-loader dependency to v6,
the issue raises again and I have no idea why it happens.
I think disabling `esModule` is the only workaround we can use...

[1]: https://github.com/vuejs/vue-style-loader/issues/50

Signed-off-by: Chawye Hsu <su+git@chawyehsu.com>